+++ src/pbd/realarray.c
@@ -23,7 +23,7 @@ static pbd_double_min_max get_min_max(const pbd_real_array* s) {
     if (s->size) {
         min_max.min = DBL_MAX;
         min_max.max = DBL_MIN;
-        for (int i = 0; i < s->size; ++i) {
+        for (size_t i = 0; i < s->size; ++i) {
             min_max.min = fmin(min_max.min, s->values[i]);
             min_max.max = fmax(min_max.max, s->values[i]);
         }
@@ -59,7 +59,7 @@ static int real_array_to_buffer(pbd_conf conf, const pbd_element* e,
     pbd_write_array_size(*buffer, s->size, sizeof_array_size);
     for (size_t i = 0; i < s->size; ++i) {
         if (sizeof_value == sizeof(float)) {
-            float value = s->values[i];  
+            float value = (float) s->values[i];  
             memcpy(*buffer + SIZEOF_TYPE_ID + sizeof_array_size + sizeof_value * i, 
                     &value, sizeof_value);
         } else {
@@ -85,7 +85,7 @@ static int real_array_from_buffer( pbd_conf conf, struct pbd_element* e,
     size_t sizeof_array_size = pbd_sizeof_array_size_by_type(type_id);
     uint32_t size = pbd_read_array_size(buffer, sizeof_array_size);
     bool little_endian = pbd_is_little_endian();
-    for (int i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         double value;
         if (sizeof_value == sizeof(float)) {
             float tmp_value;
